MyTeamPage->MyTeamModal 변경
