+++ src/views/myTeam/MyTeamMemberPage.vue
@@ -1,30 +1,25 @@
 <template>
     <div>
         <v-container class="px-15">
+
             <div class="d-flex">
                 <v-subheader>개인프로필</v-subheader>
-                <v-btn class="ml-auto" 
-                       color="black white--text" small
-                       to="/myTeamsProfile"
-                       :teamSeq="teamSeq">프로필 수정</v-btn>
-                <v-btn class="" 
-                       color="black white--text" small
-                       to="/myTeam/info"
-                       :teamSeq="teamSeq">팀정보 수정</v-btn>
+                <v-btn class="ml-auto" color="black white--text" small to="/myTeamsProfile" :teamSeq="pTeamSeq">프로필 수정</v-btn>
+                <v-btn class="" color="black white--text" small to="/myTeam/info" :teamSeq="pTeamSeq">팀정보 수정</v-btn>
             </div>
-            <MyProfile :data="profile"/>
+            <MyProfile :data="profile" />
 
             <v-subheader>운영진</v-subheader>
             <div v-for="(manager, index) in managerList" v-bind:key="index">
-                <MyManager v-bind:data="manager"/>
+                <MyManager v-bind:data="manager" />
             </div>
-        
+
             <div class="d-flex">
                 <v-subheader>팀원 목록</v-subheader>
                 <v-btn class="ml-auto" color="black white--text" small>팀원 추가</v-btn>
             </div>
-            <div v-for="(member, index) in memberList" v-bind:key="index">                    
-                <MyMember v-bind:data="member"/>
+            <div v-for="(member, index) in memberList" v-bind:key="index">
+                <MyMember v-bind:data="member" />
             </div>
 
         </v-container>
@@ -66,17 +61,6 @@
                 } catch(e) {
                     console.log(e);
                 }
-                // .catch(function(error) {
-                //     if(error.response) {
-                //         console.log(error.response.data);
-                //         console.log(error.response.status);
-                //         console.log(error.response.headers);
-                //     } else if(error.request) {
-                //         console.log(error.request);
-                //     } else {
-                //         console.log(error.message);
-                //     }
-                // })
             },
             async getListManager() {
                 try {
